Make movement smoother.
When you finish moving the length of one tile, we used to go to an idle state, which would run the
next frame. The problem is that, if the user had not let go of the direction button, the idle frame
looks like a stutter.

Now, when you finish moving the length of one tile, go to an idle state, but also run the idle
state's update routine immediately. That routine will decide whether to stay in the idle state or
go back to the moving state. This keeps movement seamless.